+++ Plutarch/PLam.hs
@@ -1,0 +1,75 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE DefaultSignatures #-}
+{-# LANGUAGE FunctionalDependencies #-}
+{-# LANGUAGE QuantifiedConstraints #-}
+{-# LANGUAGE UndecidableInstances #-}
+
+module Plutarch.PLam (
+  (PI.:-->),
+  PI.perror,
+  PI.pforce,
+  PI.phoistAcyclic,
+  PI.plam',
+  plam,
+  PI.plet,
+  PI.Term,
+  PI.TermCont (..),
+  PI.S,
+  PI.PType,
+  (#$),
+  (#),
+  pinl,
+) where
+
+import Data.Kind (Type)
+import Plutarch.Internal (PType, S, Term, papp, plam', (:-->))
+import qualified Plutarch.Internal as PI
+
+{- |
+  High precedence infixl synonym of 'papp', to be used like
+  function juxtaposition. e.g.:
+
+  >>> f # x # y
+  f x y
+-}
+(#) :: Term s (a :--> b) -> Term s a -> Term s b
+(#) = papp
+
+infixl 8 #
+
+{- |
+  Low precedence infixr synonym of 'papp', to be used like
+  '$', in combination with '#'. e.g.:
+
+  >>> f # x #$ g # y # z
+  f x (g y z)
+-}
+(#$) :: Term s (a :--> b) -> Term s a -> Term s b
+(#$) = papp
+
+infixr 0 #$
+
+{- $plam
+ Lambda abstraction.
+
+ The 'PLamN' constraint allows
+ currying to work as expected for any number of arguments.
+
+ > id :: Term s (a :--> a)
+ > id = plam (\x -> x)
+
+ > const :: Term s (a :--> b :-> a)
+ > const = plam (\x y -> x)
+-}
+
+class PLamN (a :: Type) (b :: PType) (s :: S) | a -> b, s b -> a where
+  plam :: forall c. (Term s c -> a) -> Term s (c :--> b)
+
+instance (a' ~ Term s a) => PLamN a' a s where
+  plam = plam'
+
+instance {-# OVERLAPPING #-} (a' ~ Term s a, PLamN b' b s) => PLamN (a' -> b') (a :--> b) s where
+  plam f = plam' $ \x -> plam (f x)
+
+pinl :: Term s a -> (Term s a -> Term s b) -> Term s b
+pinl v f = f v
